change Message to Alert, simplify redusers
