+++ app/src/components/Alerts/AlertsActions.js
@@ -21,6 +21,15 @@ const messageView = view => ({
   view: view
 });
 
+export const contactInfo = data => dispatch =>
+  request()
+    .post(`${API.URL}/api/support`)
+    .send(data)
+    .end((err, res) => {
+      dispatch(messageAlert(JSON.parse(res.text).message)) &&
+      dispatch(messageView(JSON.parse(res.text).view)) && dispatch(showAlert(true));
+    });
+
 export const signupUser = (email, password) => {
   let data = {
     email: email,
