+++ app/src/containers/Home/Home.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Grid, Row, Col, PageHeader } from 'react-bootstrap';
-import { Login, Message } from 'src/components';
+import { Login, Alerts } from 'src/components';
 import './Home.scss';
 
 export default class Home extends React.Component {
@@ -8,7 +8,7 @@ export default class Home extends React.Component {
   render() {
     return (
       <Grid>
-        <Message />
+        <Alerts />
         <Row className='show-grid'>
           <Col sm={12} md={8}>
             <PageHeader className='text-center'>Da-Mi</PageHeader>
